% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noun_table.R
\name{get_noun_table}
\alias{get_noun_table}
\title{Get Noun Table}
\usage{
get_noun_table(english_noun, greek_noun, add_vocative = FALSE)
}
\arguments{
\item{english_noun}{The English noun to lookup}

\item{greek_noun}{The Greek noun to lookup}

\item{add_vocative}{Logicial. The vocative case is usually not needed.}
}
\value{
A tibble
}
\description{
Look up the Greek declension table for a given noun
}
\details{
Greek nouns take on different endings or declensions depending
on case and number. This contrasts to English which usually changes by
number only, often with the addition of an s on the end. To learn Greek you
need to learn all the declensions.
}
